requirements.txt
# python=3.10.12
fastapi==0.111.0
uvicorn==0.30.1
pytz==2024.1
python-multipart==0.0.9
google-cloud-vision==3.7.2
pytesseract==0.3.10
pymongo==4.7.3
python-dateutil==2.9.0.post0
opencv-python==4.10.0.84
motor==3.4.0
python-jose==3.3.0

passlib[bcrypt]
tqdm
ipython
requests
Pillow